Create a Streamlit prototype for the shopper use case
A simple UI was created with Streamlit to showcase the use case of
providing a virtual personal shopper. The virtual assistent helps pick
out a few combinations of clothing for the user to wear based on a
conversation. In addition, the virtual assistent is collecting customer
insights to create a more personalized experience.